Create patch#0.0.2
This patch adds the framework for a menu. The ui are mainly placeholders for the menu and I plan to change it later. The menu doesn't have anything functional on it yet other than the ability to switch pages and some other things. This patch lays down the framework for future updates.