Settings Toggle Framework & Optimizations!
This patch adds the basics of a ui library to the settings. It will change further because it is the way I originally set up the tabs between them are not done well though still functional. This patch adds a few things in the settings and the basics for a ui library included in the settings (Not a whole UI lib but rather UI lib elements).